ManifoldDiffEq

The package ManifoldDiffEq aims to provide a library of differential equation solvers on manifolds. The library is built on top of Manifolds.jl and follows the interface of OrdinaryDiffEq.jl.

The code below demonstrates usage of ManifoldDiffEq to solve a simple equation and visualize the results.

Methods implemented in this library are described for example in[HLW10].

ManifoldDiffEq.ManifoldODEProblem — Type

ManifoldODEProblem

A general problem for ODE problems on Riemannian manifolds.

Fields

f the tangent vector field f(u,p,t)

u0 the initial condition

tspan time interval for the solution

p constant parameters for f`

kwargs A callback to be applied to every solver which uses the problem.

problem_type type of problem

manifold the manifold the vector field is defined on

Literature