Error measures

ManoptExamples.mean_average_error — Method

mean_average_error(M,x,y)

Compute the (mean) squared error between the two points x and y on the PowerManifold manifold M.

ManoptExamples.mean_squared_error — Method

mean_squared_error(M, p, q)

Compute the (mean) squared error between the two points p and q on the (power) manifold M.