Grassmannian manifold

Manifolds.Grassmann — Type

Grassmann{n,k,𝔽} <: AbstractEmbeddedManifold{𝔽,DefaultIsometricEmbeddingType}

The Grassmann manifold $\operatorname{Gr}(n,k)$ consists of all subspaces spanned by $k$ linear independent vectors $𝔽^n$, where $𝔽 ∈ \{ℝ, ℂ\}$ is either the real- (or complex-) valued vectors. This yields all $k$-dimensional subspaces of $ℝ^n$ for the real-valued case and all $2k$-dimensional subspaces of $ℂ^n$ for the second.

The manifold can be represented as

\[\operatorname{Gr}(n,k) := \bigl\{ \operatorname{span}(p) : p ∈ 𝔽^{n × k}, p^\mathrm{H}p = I_k\},\]

where $\cdot^{\mathrm{H}}$ denotes the complex conjugate transpose or Hermitian and $I_k$ is the $k × k$ identity matrix. This means, that the columns of $p$ form an unitary basis of the subspace, that is a point on $\operatorname{Gr}(n,k)$, and hence the subspace can actually be represented by a whole equivalence class of representers. Another interpretation is, that

\[\operatorname{Gr}(n,k) = \operatorname{St}(n,k) / \operatorname{O}(k),\]

i.e the Grassmann manifold is the quotient of the Stiefel manifold and the orthogonal group $\operatorname{O}(k)$ of orthogonal $k × k$ matrices.

The tangent space at a point (subspace) $x$ is given by

\[T_x\mathrm{Gr}(n,k) = \bigl\{ X ∈ 𝔽^{n × k} : X^{\mathrm{H}}p + p^{\mathrm{H}}X = 0_{k} \bigr\},\]

where $0_{k}$ denotes the $k × k$ zero matrix.

Note that a point $p ∈ \operatorname{Gr}(n,k)$ might be represented by different matrices (i.e. matrices with unitary column vectors that span the same subspace). Different representations of $p$ also lead to different representation matrices for the tangent space $T_p\mathrm{Gr}(n,k)$

The manifold is named after Hermann G. Graßmann (1809-1877).

Constructor

Grassmann(n,k,field=ℝ)

Generate the Grassmann manifold $\operatorname{Gr}(n,k)$, where the real-valued case field = ℝ is the default.

Base.exp — Method

exp(M::Grassmann, p, X)

Compute the exponential map on the GrassmannM$= \mathrm{Gr}(n,k)$ starting in p with tangent vector (direction) X. Let $X = USV$ denote the SVD decomposition of $X$. Then the exponential map is written using

\[z = p V\cos(S)V^\mathrm{H} + U\sin(S)V^\mathrm{H},\]

where $\cdot^{\mathrm{H}}$ denotes the complex conjugate transposed or Hermitian and the cosine and sine are applied element wise to the diagonal entries of $S$. A final QR decomposition $z=QR$ is performed for numerical stability reasons, yielding the result as

\[\exp_p X = Q.\]

Base.log — Method

log(M::Grassmann, p, q)

Compute the logarithmic map on the GrassmannM$ = \mathcal M=\mathrm{Gr}(n,k)$, i.e. the tangent vector X whose corresponding geodesic starting from p reaches q after time 1 on M. The formula reads

\[\log_p q = V\cdot \operatorname{atan}(S) \cdot U^\mathrm{H},\]

where $\cdot^{\mathrm{H}}$ denotes the complex conjugate transposed or Hermitian. The matrices $U$ and $V$ are the unitary matrices, and $S$ is the diagonal matrix containing the singular values of the SVD-decomposition

\[USV = (q^\mathrm{H}p)^{-1} ( q^\mathrm{H} - q^\mathrm{H}pp^\mathrm{H}).\]

In this formula the $\operatorname{atan}$ is meant elementwise.

Manifolds.uniform_distribution — Method

uniform_distribution(M::Grassmann{n,k,ℝ}, p)

Uniform distribution on given (real-valued) GrassmannM. Specifically, this is the normalized Haar measure on M. Generated points will be of similar type as p.

The implementation is based on Section 2.5.1 in [Chikuse2003]; see also Theorem 2.2.2(iii) in [Chikuse2003].

ManifoldsBase.check_manifold_point — Method

check_manifold_point(M::Grassmann{n,k,𝔽}, p)

Check whether p is representing a point on the GrassmannM, i.e. its a n-by-k matrix of unitary column vectors and of correct eltype with respect to 𝔽.

ManifoldsBase.check_tangent_vector — Method

check_tangent_vector(M::Grassmann{n,k,𝔽}, p, X; check_base_point = true, kwargs...)

Check whether X is a tangent vector in the tangent space of p on the GrassmannM, i.e. that X is of size and type as well as that

\[ p^{\mathrm{H}}X + X^{\mathrm{H}}p = 0_k,\]

where $\cdot^{\mathrm{H}}$ denotes the complex conjugate transpose or Hermitian and $0_k$ denotes the $k × k$ zero natrix. The optional parameter check_base_point indicates, whether to call check_manifold_point for p.

ManifoldsBase.distance — Method

distance(M::Grassmann, p, q)

Compute the Riemannian distance on Grassmann manifold M$= \mathrm{Gr}(n,k)$.

Let $USV = p^\mathrm{H}q$ denote the SVD decomposition of $p^\mathrm{H}q$, where $\cdot^{\mathrm{H}}$ denotes the complex conjugate transposed or Hermitian. Then the distance is given by

\[d_{\mathrm{Gr}(n,k)}(p,q) = \operatorname{norm}(\operatorname{Re}(b)).\]

where

\[b_{i}=\begin{cases} 0 & \text{if} \; S_i ≥ 1\\ \arccos(S_i) & \, \text{if} \; S_i<1. \end{cases}\]

ManifoldsBase.injectivity_radius — Method

injectivity_radius(M::Grassmann)
injectivity_radius(M::Grassmann, p)

Return the injectivity radius on the GrassmannM, which is $\frac{π}{2}$.

ManifoldsBase.inner — Method

inner(M::Grassmann, p, X, Y)

Compute the inner product for two tangent vectors X, Y from the tangent space of p on the Grassmann manifold M. The formula reads

\[g_p(X,Y) = \operatorname{tr}(X^{\mathrm{H}}Y),\]

where $\cdot^{\mathrm{H}}$ denotes the complex conjugate transposed or Hermitian.

ManifoldsBase.inverse_retract — Method

inverse_retract(M::Grassmann, p, q, ::PolarInverseRetraction)

Compute the inverse retraction for the PolarRetraction, on the Grassmann manifold M, i.e.,

\[\operatorname{retr}_p^{-1}q = q*(p^\mathrm{H}q)^{-1} - p,\]

where $\cdot^{\mathrm{H}}$ denotes the complex conjugate transposed or Hermitian.

ManifoldsBase.inverse_retract — Method

inverse_retract(M, p, q, ::QRInverseRetraction)

Compute the inverse retraction for the QRRetraction, on the Grassmann manifold M, i.e.,

\[\operatorname{retr}_p^{-1}q = q(p^\mathrm{H}q)^{-1} - p,\]

where $\cdot^{\mathrm{H}}$ denotes the complex conjugate transposed or Hermitian.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::Grassmann)

Return the dimension of the Grassmann(n,k,𝔽) manifold M, i.e.

\[\dim \operatorname{Gr}(n,k) = k(n-k) \dim_ℝ 𝔽,\]

where $\dim_ℝ 𝔽$ is the real_dimension of 𝔽.

ManifoldsBase.project — Method

project(M::Grassmann, p, X)

Project the n-by-kX onto the tangent space of p on the GrassmannM, which is computed by

\[\operatorname{proj_p}(X) = X - pp^{\mathrm{H}}X,\]

where $\cdot^{\mathrm{H}}$ denotes the complex conjugate transposed or Hermitian.

ManifoldsBase.representation_size — Method

representation_size(M::Grassmann{n,k})

Return the represenation size or matrix dimension of a point on the GrassmannM, i.e. $(n,k)$ for both the real-valued and the complex value case.

ManifoldsBase.retract — Method

retract(M::Grassmann, p, X, ::PolarRetraction)

Compute the SVD-based retraction PolarRetraction on the GrassmannM. With $USV = p + X$ the retraction reads

\[\operatorname{retr}_p X = UV^\mathrm{H},\]

where $\cdot^{\mathrm{H}}$ denotes the complex conjugate transposed or Hermitian.

ManifoldsBase.retract — Method

retract(M::Grassmann, p, X, ::QRRetraction )

Compute the QR-based retraction QRRetraction on the GrassmannM. With $QR = p + X$ the retraction reads

\[\operatorname{retr}_p X = QD,\]

where D is a $m × n$ matrix with

\[D = \operatorname{diag}( \operatorname{sgn}(R_{ii}+0,5)_{i=1}^n ).\]

ManifoldsBase.vector_transport_to — Method

vector_transport_to(M::Grassmann,p,X,q,::ProjectionTransport)

compute the projection based transport on the GrassmannM by interpreting X from the tangent space at p as a point in the embedding and projecting it onto the tangent space at q.

ManifoldsBase.zero_tangent_vector — Method

zero_tangent_vector(M::Grassmann, p)

Return the zero tangent vector from the tangent space at p on the GrassmannM, which is given by a zero matrix the same size as p.

Statistics.mean — Method

mean(
    M::Grassmann,
    x::AbstractVector,
    [w::AbstractWeights,]
    method = GeodesicInterpolationWithinRadius(π/4);
    kwargs...,
)

Compute the Riemannian mean of x using GeodesicInterpolationWithinRadius.

Chikuse2003

Y. Chikuse: "Statistics on Special Manifolds", Springer New York, 2003, doi: 10.1007/978-0-387-21540-2.