The Symmetric Matrices mathrmSym(n)

The $n\times n$ symmetric matrices $\mathrm{Sym}(n)$ embedded in $\mathbb R^{n\times n}$

The manifold of symmetric matrices $\mathcal{Sym}(n)$ posesses the following instances of the abstract types Manifold, MPoint, and TVector.

Manopt.SymPoint — Type.

SymPoint <: MPoint

A point $x$ on the manifold $\mathcal M = \mathrm{Sym}(n)$ of $n\times n$ symmetric matrices, represented in the redundant way of a symmetric matrix (instead of storing just the upper half).

source

Manopt.SymTVector — Type.

SymTVector <: TVector

A tangent vector $\xi$ in $T_x\mathcal M$ of a symmetric matrix $x\in\mathcal M$.

source

Manopt.Symmetric — Type.

Symmetric <: Manifold

The manifold $\mathcal M = \mathrm{Sym}(n)$, where $\mathrm{Sym}(n) = \{ x \in \mathbb R^{n\times n} | x = x^\mathrm{T} \}$, $n\in\mathbb N$, denotes the manifold of symmetric matrices equipped with the trace inner product and its induced Forbenius norm.

Abbreviation

Sym or Sym(n), respectively.

Constructor

Symmetric(n)

generates the manifold of n-by-n symmetric matrices.

source

Functions

Base.exp — Function.

exp(M,x,ξ[, t=1.0])

compute the exponential map on the Symmetric manifold M given a SymPointx and a SymTVectorξ, as well as an optional scaling factor t. The exponential map is given by

\[\exp_{x}ξ = x+ξ.\]

source

Base.log — Method.

log(M,x,y)

compute the logarithmic map for two SymPointx,y on the SymmetricM, which is given by $\log_xy = y-x$.

source

LinearAlgebra.dot — Method.

dot(M,x,ξ,ν)

inner product of two SymTVectors ξ,ν lying in the tangent space of the SymPointx on the Symmetric manifold M.

source

LinearAlgebra.norm — Method.

norm(M,x,ξ)

computes the norm of the SymTVectorξ in the tangent space of the SymPointx on the SymmetricM embedded in the Euclidean space, i.e. by its Frobenius norm.

source

Manopt.distance — Method.

distance(M,x,y)

distance of two SymPoints x,y on the Symmetric manifold M` inherited from embedding them in $\mathbb R^{n\times n}$, i.e. use the Frobenious norm of the difference.

source

Manopt.manifoldDimension — Method.

manifoldDimension(x)

returns the manifold dimension the SymPointx belongs to.

source

Manopt.manifoldDimension — Method.

manifoldDimension(M)

returns the manifold dimension of the Symmetric manifold M.

source

Manopt.parallelTransport — Method.

parallelTransport(M,x,y,ξ)

compute the parallel transport of a SymTVectorξ from the tangent space at the SymPointx to the SymPointy on the Symmetric manifold M. Since the metric is inherited from the embedding space, it is just the identity.

source

Manopt.typicalDistance — Method.

typicalDistance(M)

returns the typical distance on the Symmetric manifold M, i.e. $\sqrt{n}$.

source

Manopt.validateMPoint — Method.

validateMPoint(M,x)

validate, that the SymPointx is a valid point on the Symmetric manifold M, i.e. that its dimensions are correct and that the matrix is symmetric.

source

Manopt.validateTVector — Method.

validateTVector(M,x,ξ)

validate, that the SymTVector is a valid tangent vector to the SymPointx on the Symmetric manifold M, i.e. that its dimensions are correct and that the matrix is symmetric.

source

Manopt.zeroTVector — Method.

ξ = zeroTVector(M,x)

returns a zero vector in the tangent space $T_x\mathcal M$ of the SymPointx on the Symmetric manifold M.

source