Plans for solvers

For any optimisation performed in Manopt.jl we need information about both the optimisation task or “problem” at hand as well as the solver and all its parameters. This together is called a plan in Manopt.jl and it consists of two data structures:

The Manopt Problem describes all static data of our task, most prominently the manifold and the objective.

The Solver State describes all varying data and parameters for the solver we aim to use. This also means that each solver has its own data structure for the state.

By splitting these two parts, we can use one problem and solve it using different solvers.

Still there might be the need to set certain parameters within any of these structures. For that there is

Manopt.set_manopt_parameter! — Function

set_manopt_parameter!(f, element::Symbol , args...)

For any f and a Symbole we dispatch on its value so by default, to set some args... in f or one of uts sub elements.

set_manopt_parameter!(amo::AbstractManifoldObjective, element::Symbol, args...)

Set a certain args... from the AbstractManifoldObjectiveamo to value. This function should dispatch onVal(element)`.

Currently supported

:Cost passes to the get_cost_function

:Gradient passes to the get_gradient_function

set_manopt_parameter!(ams::AbstractManoptProblem, element::Symbol, field::Symbol , value)

Set a certain field/element from the AbstractManoptProblemams to value. This function should dispatch onVal(element)`.

By default this passes on to the inner objective, see set_manopt_parameter!

set_manopt_parameter!(ams::DebugSolverState, ::Val{:Debug}, args...)

Set certain values specified by args... into the elements of the debugDictionary

set_manopt_parameter!(ams::AbstractManoptSolverState, element::Symbol, args...)

Set a certain field or semantic element from the AbstractManoptSolverStateams to value. This function passes to Val(element) and specific setters should dispatch on Val{element}.

By default, this function just does nothing.

set_manopt_parameter!(ams::DebugSolverState, ::Val{:SubProblem}, args...)

Set certain values specified by args... to the sub problem.

set_manopt_parameter!(ams::DebugSolverState, ::Val{:SubState}, args...)

Set certain values specified by args... to the sub state.

Manopt.get_manopt_parameter — Function

get_manopt_parameter(ams::AbstractManoptSolverState, element::Symbol, args...)

Obtain a certain field or semantic element from the AbstractManoptSolverStateams. This function passes to Val(element) and specific setters should dispatch on Val{element}.

Manopt.status_summary — Function

status_summary(e)

Return a string reporting about the current status of e, where e is a type from Manopt, e.g. an AbstractManoptSolverStates.

This method is similar to show but just returns a string. It might also be more verbose in explaining, or hide internal information.

Where the following Symbols are used

The following symbols are used. The column “generic” refers to a short hand that might be used – for readability if clear from context.

| Symbol | Used in | Description | generic | | :–––––- | :–––: | ;–––––––––––––––––––––––––––– | :––– | | :active | DebugWhenActive | activity of the debug action stored within | | | :Basepoint | TangentSpace | the point the tangent space is at | :p | | :Cost | generic |the cost function (e.g. within an objective, as pass down) | | | :Debug | DebugSolverState | the stored debugDictionary | | | :Gradient | generic |the gradient function (e.g. within an objective, as pass down) | | | :Iterate | generic | the (current) iterate – similar to set_iterate! – within a state | | | :Manifold | generic |the manifold (e.g. within a problem, as pass down) | | | :Objective | generic | the objective (e.g. within a problem, as pass down) | | | :SubProblem | generic | the sub problem (e.g. within a state, as pass down) | | | :SubState | generic | the sub state (e.g. within a state, as pass down) | | | :λ | ProximalDCCost, ProximalDCGrad | set the proximal parameter within the proximal sub objective elements | | | :p | generic | a certain point | | | :X | generic | a certain tangent vector | | | :TrustRegionRadius | TrustRegionsState | the trust region radius | :σ | | :ρ, :u | ExactPenaltyCost, ExactPenaltyGrad | Parameters within the exact penalty objetive | | | :ρ, :μ, :λ | AugmentedLagrangianCost and AugmentedLagrangianGrad | Parameters of the Lagrangian function | |