Lie group solvers

An initial value problem manifold ordinary differential equation in the Lie action formulation.

A Lie ODE on manifold $M$ is defined in terms a vector field $F: (M × P × ℝ) \to 𝔤$ where $𝔤$ is the Lie algebra of a Lie group $G$ acting on $M$, with an initial value $y_0$ and $P$ is the space of constant parameters. A solution to this problem is a curve $y\colon ℝ\to M$ such that $y(0)=y_0$ and for each $t \in [0, T]$ we have $D_t y(t) = f(y(t), p, t)\circ y(t)$, where the $\circ$ is defined as

\[X\circ m = \frac{d}{dt}\vert_{t=0} \exp(tZ)\cdot m\]

and $\cdot$ is the group action of $G$ on $M$.

The Lie group $G$ must act transitively on $M$, that is for each pair of points $p, q$ on $M$ there is an element $a \in G$ such that $a\cdot p = q$. See for example [CMO14] for details.

ManifoldDiffEq.ManifoldLieEuler — Type

ManifoldLieEuler

The manifold Lie-Euler algorithm for problems in the LieODEProblemType formulation.

ManifoldDiffEq.ManifoldLieEulerCache — Type

ManifoldLieEulerCache

Cache for ManifoldLieEuler.

ManifoldDiffEq.ManifoldLieEulerConstantCache — Type

ManifoldLieEulerConstantCache

Constant cache for ManifoldLieEuler.

ManifoldDiffEq.RKMK4 — Type

RKMK4

The Lie group variant of fourth-order Runge-Kutta algorithm for problems in the LieODEProblemType formulation, called Runge-Kutta Munthe-Kaas. The Butcher tableau is:

\[\begin{array}{c|cccc} 0 & 0 \\ \frac{1}{2} & 0 & \frac{1}{2} & 0 \\ \frac{1}{2} & \frac{1}{2} & 0 \\ 1 & 0 & 0 & 1 & 0\\ \hline & \frac{1}{6} & \frac{1}{3} & \frac{1}{6} & \frac{1}{6} \end{array}\]

For more details see [MO99].

ManifoldDiffEq.RKMK4Cache — Type

RKMK4Cache

Cache for RKMK4.

ManifoldDiffEq.RKMK4ConstantCache — Type

RKMK4ConstantCache

Constant cache for RKMK4.

ManifoldDiffEq.LieODEProblemType — Type

LieODEProblemType

An initial value problem manifold ordinary differential equation in the Lie action formulation.

A Lie ODE on manifold $M$ is defined in terms a vector field $F: (ℝ × P × M) \to 𝔤$ where $𝔤$ is the Lie algebra of a Lie group $G$ acting on $M$, with an initial value $y₀$ and $P$ is the space of constant parameters. A solution to this problem is a curve $y:ℝ\to M$ such that $y(0)=y₀$ and for each $t ∈ [0, T]$ we have $D_t y(t) = F(y(t), p, t)∘y(t)$, where the $∘$ is defined as

\[X∘m = \frac{d}{dt}\vert_{t=0} \exp(tZ)⋅m\]

and $⋅$ is the group action of $G$ on $M$.

Note

Proofs of convergence and order have several assumptions, including time-independence of $F$. Integrators may not work well if these assumptions do not hold.

ManifoldDiffEq.LieManifoldDiffEqOperator — Type

LieManifoldDiffEqOperator{T<:Number,TF} <: AbstractDiffEqOperator{T}

DiffEq operator on manifolds in the Lie group action formulation.

Literature