Product manifold

Product manifold $\mathcal M = \mathcal{M}_1 × \mathcal{M}_2 × … × \mathcal{M}_n$ of manifolds $\mathcal{M}_1, \mathcal{M}_2, …, \mathcal{M}_n$. Points on the product manifold can be constructed using ArrayPartition (from RecursiveArrayTools.jl) with canonical projections $Π_i : \mathcal{M} → \mathcal{M}_i$ for $i ∈ 1, 2, …, n$ provided by submanifold_component.

Manifolds.ProductFVectorDistribution — Type

ProductFVectorDistribution([type::VectorSpaceFiber], [x], distrs...)

Generates a random vector at point x from vector space (a fiber of a tangent bundle) of type type using the product distribution of given distributions.

Vector space type and x can be automatically inferred from distributions distrs.

Manifolds.ProductPointDistribution — Type

ProductPointDistribution(M::ProductManifold, distributions)

Product distribution on manifold M, combined from distributions.

ManifoldDiff.riemannian_Hessian — Method

Y = riemannian_Hessian(M::ProductManifold, p, G, H, X)
riemannian_Hessian!(M::ProductManifold, Y, p, G, H, X)

Compute the Riemannian Hessian $\operatorname{Hess} f(p)[X]$ given the Euclidean gradient $∇ f(\tilde p)$ in G and the Euclidean Hessian $∇^2 f(\tilde p)[\tilde X]$ in H, where $\tilde p, \tilde X$ are the representations of $p,X$ in the embedding,.

On a product manifold, this decouples and can be computed elementwise.

Manifolds.flat — Method

flat(M::ProductManifold, p, X::FVector{TangentSpaceType})

use the musical isomorphism to transform the tangent vector X from the tangent space at p on the ProductManifoldM to a cotangent vector. This can be done elementwise for every entry of X (with respect to the corresponding entry in p) separately.

Manifolds.manifold_volume — Method

manifold_volume(M::ProductManifold)

Return the volume of ProductManifoldM, i.e. product of volumes of the manifolds M is constructed from.

Manifolds.sharp — Method

sharp(M::ProductManifold, p, ξ::FVector{CotangentSpaceType})

Use the musical isomorphism to transform the cotangent vector ξ from the tangent space at p on the ProductManifoldM to a tangent vector. This can be done elementwise for every entry of ξ (and p) separately

Manifolds.volume_density — Method

volume_density(M::ProductManifold, p, X)

Return volume density on the ProductManifoldM, i.e. product of constituent volume densities.